Created the CreateRental file
